+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog]
 
 ## [Unreleased]
 
+- Service operators can reject a claim
 - Show DfE number against school in approver view of claim
 - Drop redundant full_name column from claims table
 - Updated Accessibility statement to show current issues